+++ examples/next/kitchen-sink/app/page.tsx
@@ -4,6 +4,6 @@ import ClientPage from './[...urlSegments]/client-page';
 export const revalidate = 300;
 
 export default async function Home() {
-  const data = await client.queries.page({ relativePath: 'home.mdx' });
+  const data = await client.queries.page({ relativePath: 'home.md' });
   return <ClientPage {...data} />;
 }