+++ packages/@tinacms/schema-tools/src/types/index.ts
@@ -334,7 +334,7 @@ export type PasswordField = (
     type: 'password';
   };
 
-type ToolbarOverrideType =
+export type ToolbarOverrideType =
   | 'heading'
   | 'link'
   | 'image'
@@ -345,10 +345,13 @@ type ToolbarOverrideType =
   | 'codeBlock'
   | 'bold'
   | 'italic'
+  | 'strikethrough'
+  | 'highlight'
   | 'raw'
   | 'embed'
   | 'mermaid'
-  | 'table';
+  | 'table'
+  | 'hr';
 type RichTextAst = { type: 'root'; children: Record<string, unknown>[] };
 export type RichTextField<WithNamespace extends boolean = false> = (
   | FieldGeneric<RichTextAst, undefined>