+++ packages/tinacms/src/auth/TinaCloudProvider.tsx
@@ -63,7 +63,7 @@ const AuthWallInner = ({
   const loginScreen = client.authProvider.getLoginScreen();
   if (loginStrategy === 'LoginScreen' && !loginScreen) {
     throw new Error(
-      'LoginScreen is set as the login strategy but no login screen component was provided'
+      'LoginScreen is set as the login strategy but no login screen component was provided.'
     );
   }
 
@@ -99,7 +99,7 @@ const AuthWallInner = ({
               } else {
                 setErrorMessage({
                   title: 'Access Denied:',
-                  message: 'Not Authorized To Edit',
+                  message: 'Not authorized to edit.',
                 });
                 setActiveModal('error');
               }
@@ -173,7 +173,7 @@ const AuthWallInner = ({
     }
   };
 
-  let modalTitle = 'Let’s get you editing with TinaCMS...';
+  let modalTitle = 'Let’s get you editing with TinaCMS!';
   if (
     activeModal === 'authenticate' &&
     loginStrategy === 'Redirect' &&
@@ -184,12 +184,12 @@ const AuthWallInner = ({
     activeModal === 'authenticate' &&
     loginStrategy === 'UsernamePassword'
   ) {
-    modalTitle = 'Let’s get you editing with TinaCMS...';
+    modalTitle = 'Let’s get you editing with TinaCMS!';
   } else if (activeModal === 'error') {
     if (loginStrategy === 'Redirect' && !isTinaCloud) {
       modalTitle = 'Enter into edit mode';
     } else if (loginStrategy === 'UsernamePassword') {
-      modalTitle = 'Let’s get you editing with TinaCMS...';
+      modalTitle = 'Let’s get you editing with TinaCMS!';
     }
   }
 
@@ -205,7 +205,7 @@ const AuthWallInner = ({
                 alt='Tina the Llama playing a large orange key like a guitar.'
                 width={816}
                 height={816}
-                style={{ maxWidth: '100%', margin: '0 auto', display: 'block' }}
+                style={{ width: '16rem', margin: '0 auto', display: 'block' }}
               />
             ) : (
               'When you save, changes will be saved to the local filesystem.'