+++ packages/tinacms-gitprovider-github/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tinacms-gitprovider-github
 
+## 4.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@2.0.15
+
+## 4.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@2.0.14
+
 ## 4.1.0
 
 ### Minor Changes